Existing code shorter, adds on-screen sharp keys, CSS file redundant
The sharp keys are on the top of the video player, and are played by clicking them. There's no keyboard shortcuts because the YouTube player has none, but I may add support via the letter keys in the future, or even all possible YouTube player controls. These and the piano may be split into modes, even into different files, if so maybe into one later.